chore(generic-binding): move to package addons
